requirements.txt
datadiff
flake8
flake8-bugbear
isort
matplotlib
numpy
pre-commit
pyyaml
tensorflow==v1.14
tqdm
